Get local data for bar chart instead of firebase
